portfolio/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile
#
astroid==2.8.5
    # via pylint
attrs==21.2.0
    # via
    #   automat
    #   service-identity
    #   twisted
autobahn==21.2.1
    # via binance-connector
automat==20.2.0
    # via twisted
binance-connector==1.7.0
    # via -r requirements.in
black==21.10b0
    # via -r requirements.in
certifi==2021.10.8
    # via requests
cffi==1.15.0
    # via cryptography
chardet==4.0.0
    # via requests
click==8.0.3
    # via black
colorama==0.4.4
    # via
    #   click
    #   pylint
constantly==15.1.0
    # via twisted
cryptography==35.0.0
    # via
    #   autobahn
    #   pyopenssl
    #   service-identity
hyperlink==21.0.0
    # via
    #   autobahn
    #   twisted
idna==2.10
    # via
    #   hyperlink
    #   requests
incremental==21.3.0
    # via twisted
isort==5.10.1
    # via pylint
krakenex==2.1.0
    # via -r requirements.in
lazy-object-proxy==1.6.0
    # via astroid
mccabe==0.6.1
    # via pylint
mypy-extensions==0.4.3
    # via black
pathspec==0.9.0
    # via black
platformdirs==2.4.0
    # via
    #   black
    #   pylint
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   service-identity
pyasn1-modules==0.2.8
    # via service-identity
pycoingecko==2.2.0
    # via -r requirements.in
pycparser==2.21
    # via cffi
pylint==2.11.1
    # via -r requirements.in
pyopenssl==19.0.0
    # via binance-connector
python-dotenv==0.19.2
    # via -r requirements.in
python-kucoin==2.1.3
    # via -r requirements.in
regex==2021.11.10
    # via black
requests==2.25.1
    # via
    #   binance-connector
    #   krakenex
    #   pycoingecko
    #   python-kucoin
service-identity==21.1.0
    # via binance-connector
six==1.16.0
    # via
    #   automat
    #   pyopenssl
    #   service-identity
toml==0.10.2
    # via pylint
tomli==1.2.2
    # via black
twisted==21.2.0
    # via binance-connector
twisted-iocpsupport==1.0.2
    # via twisted
txaio==21.2.1
    # via autobahn
typing-extensions==3.10.0.2
    # via black
urllib3==1.26.7
    # via requests
websockets==10.1
    # via python-kucoin
wrapt==1.13.3
    # via astroid
zope.interface==5.4.0
    # via twisted

# The following packages are considered to be unsafe in a requirements file:
# setuptools
